Updated employee api; Makes sure target employee is not a manager before  getting deleted

+++ app/api/employees/delete/route.ts
@@ -1,5 +1,6 @@
 import Employee from "@/models/employeeModel";
 import { connectDB } from "@/lib/database";
+import Manager from "@/models/managerModel";
 
 export const DELETE = async (request: Request) => {
   const { id } = await request.json();
@@ -11,9 +12,19 @@ export const DELETE = async (request: Request) => {
   await connectDB();
   const employee = await Employee.findById(id).exec();
 
+  const manager = await Manager.findOne({ employee: employee._id });
+
   if (!employee) {
     return new Response("Employee not found", { status: 400 });
   }
+
+  if (manager)
+    return new Response(
+      "Make sure to demote employee from manager status first",
+      {
+        status: 400,
+      }
+    );
 
   await employee.deleteOne();
